gzip frequent object creation can lead to memory leaks
