+++ implant/sliver/encoders/gzip.go
@@ -21,6 +21,7 @@ package encoders
 import (
 	"bytes"
 	"compress/gzip"
+	"sync"
 )
 
 // GzipEncoderID - EncoderID
@@ -29,12 +30,44 @@ const GzipEncoderID = 49
 // Gzip - Gzip compression encoder
 type Gzip struct{}
 
+var gzipWriterPools = &sync.Pool{}
+
+func init() {
+	gzipWriterPools = &sync.Pool{
+		New: func() interface{} {
+			w, _ := gzip.NewWriterLevel(nil, gzip.BestSpeed)
+			return w
+		},
+	}
+}
+
+// GzipBuf - Gzip a buffer
+func GzipBuf(data []byte) []byte {
+	var buf bytes.Buffer
+	gzipWriter := gzipWriterPools.Get().(*gzip.Writer)
+	gzipWriter.Reset(&buf)
+	gzipWriter.Write(data)
+	gzipWriter.Close()
+	gzipWriterPools.Put(gzipWriter)
+	return buf.Bytes()
+}
+
+// GunzipBuf - Gunzip a buffer
+func GunzipBuf(data []byte) []byte {
+	zip, _ := gzip.NewReader(bytes.NewBuffer(data))
+	var buf bytes.Buffer
+	buf.ReadFrom(zip)
+	return buf.Bytes()
+}
+
 // Encode - Compress data with gzip
 func (g Gzip) Encode(data []byte) []byte {
 	var buf bytes.Buffer
-	gzipWriter, _ := gzip.NewWriterLevel(&buf, gzip.BestSpeed)
+	gzipWriter := gzipWriterPools.Get().(*gzip.Writer)
+	gzipWriter.Reset(&buf)
 	gzipWriter.Write(data)
 	gzipWriter.Close()
+	gzipWriterPools.Put(gzipWriter)
 	return buf.Bytes()
 }
