+++ util/encoders/gzip_test.go
@@ -2,9 +2,10 @@ package encoders
 
 import (
 	"bytes"
-	"testing"
-
+	"crypto/rand"
 	implantEncoders "github.com/bishopfox/sliver/implant/sliver/encoders"
+	insecureRand "math/rand"
+	"testing"
 )
 
 /*
@@ -58,5 +59,22 @@ func TestGzip(t *testing.T) {
 		t.Logf("output1 = %#v", output)
 		t.Logf("output2 = %#v", output2)
 		t.Errorf("server/implant outputs do not match returned\n%#v != %#v", sample, data)
+	}
+}
+
+func randomDataRandomSize(maxSize int) []byte {
+	buf := make([]byte, insecureRand.Intn(maxSize))
+	rand.Read(buf)
+	return buf
+}
+
+func TestGzipGunzip(t *testing.T) {
+	for i := 0; i < 100; i++ {
+		data := randomDataRandomSize(8192)
+		gzipData := GzipBuf(data)
+		gunzipData := GunzipBuf(gzipData)
+		if !bytes.Equal(data, gunzipData) {
+			t.Fatalf("Data does not match")
+		}
 	}
 }
