+++ util/files.go
@@ -30,23 +30,6 @@ import (
 	"path/filepath"
 )
 
-// GzipBuf - Gzip a buffer
-func GzipBuf(data []byte) []byte {
-	var buf bytes.Buffer
-	zip := gzip.NewWriter(&buf)
-	zip.Write(data)
-	zip.Close()
-	return buf.Bytes()
-}
-
-// GunzipBuf - Gunzip a buffer
-func GunzipBuf(data []byte) []byte {
-	zip, _ := gzip.NewReader(bytes.NewBuffer(data))
-	var buf bytes.Buffer
-	buf.ReadFrom(zip)
-	return buf.Bytes()
-}
-
 // DeflateBuf - Deflate a buffer using BestCompression (9)
 func DeflateBuf(data []byte) []byte {
 	var buf bytes.Buffer
